backend/pyproject.toml
[tool.poetry]
name = "vla-api-alt"
version = "0.1.0"
description = ""
authors = ["pelpuo <epkayang@gmail.com>"]
readme = "README.md"
packages = [{include = "vla_api_alt"}]

[tool.poetry.dependencies]
python = "^3.8"
anyio = "3.6.2"
certifi = "2022.12.7"
charset-normalizer = "3.0.1"
click = "8.1.3"
colorama = "0.4.6"
fastapi = "0.89.1"
filelock = "3.9.0"
fuzzywuzzy = "0.18.0"
h11 = "0.14.0"
huggingface-hub = "0.12.0"
idna = "3.4"
levenshtein = "0.20.9"
numpy = "1.24.1"
packaging = "23.0"
pydantic = "1.10.4"
pypdf2 = "2.12.1"
python-levenshtein = "0.20.9"
python-multipart = "0.0.5"
pyyaml = "6.0"
rapidfuzz = "2.13.7"
regex = "2022.10.31"
requests = "2.28.2"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.22.0"
tokenizers = "0.13.2"
tqdm = "4.64.1"
transformers = "4.26.0"
typing-extensions = "4.4.0"
urllib3 = "1.26.14"
uvicorn = "0.20.0"
flair = "^0.12.2"
dateparser = "^1.1.8"
pypdfium2 = {version= "*", python = "^3.11"}
python-doctr = "^0.6.0"
torchvision = "^0.15.1"
torch = "^2.0.0"
jellyfish = "^0.11.2"
ocrmypdf = "^14.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
